move test for identity update to spec files
